Fix a warning about comparison of signed and unsigned value.
git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@4477 941e2c2b-deac-454f-805a-451daa25f33c

+++ libhfst/src/implementations/HfstTransitionGraph.h
@@ -3356,7 +3356,7 @@
                }
              // if there was previous distance defined for 'state', erase it, if needed
              int previous_distance = distance_of_state.at(state);
-             if (previous_distance != -1 && previous_distance != distance && overwrite)
+             if (previous_distance != -1 && previous_distance != (int)distance && overwrite)
                {
                  states_at_distance.at(previous_distance).erase(state);
                }
